requirements.txt
fastapi==0.101.0
uvicorn[standard]==0.22.0
asyncpg==0.27.0
python-dotenv==1.0.0